:bug: Fix: Studies 앱 블랙리스트 테스트와 가입 신청 테스트 파일 분리
1. 원인
1.1. 블랙리스트 테스트와 가입 신청 테스트가 파일 하나로 작성됨
- 명시적으로 확인이 어려움

2. 수정
2.1. 블랙리스트 테스트 파일과 가입 신청 테스트 파일을 각각 작성하였음.
